Revert "Merged SeasonalStatistics.java into TrackStatistics.java"
This reverts commit 4373b10f330c4682f0e43134aa59795569f76d09.
